reorganize LogicUtils
