+++ src/orange/orangeinp/detail/LogicUtils.hh
@@ -2,7 +2,7 @@
 // Copyright Celeritas contributors: see top-level COPYRIGHT file for details
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file orange/orangeinp/detail/BuildLogicUtils.hh
+//! \file orange/orangeinp/detail/LogicUtils.hh
 //---------------------------------------------------------------------------//
 #pragma once
 
@@ -23,6 +23,106 @@ namespace orangeinp
 {
 namespace detail
 {
+
+//---------------------------------------------------------------------------//
+/*!
+ * A helper class for keeping track of the operand type of a sub-expression.
+ */
+struct Operand
+{
+    logic::OperatorToken expr_type;
+    std::vector<logic_int> expr;
+};
+
+//---------------------------------------------------------------------------//
+/*!
+ * Convert a postfix logic expression to an infix expression.
+ *
+ * The \c InfixEvaluator will short-circuit evaluation of operands based
+ * on parenthesis depth. Minimizing that depth in the expression
+ * will allow to short-circuit more efficiently.
+ */
+inline std::vector<logic_int> convert_to_infix(Span<logic_int> postfix)
+{
+    CELER_EXPECT(!postfix.empty());
+
+    std::vector<Operand> infix;
+    infix.reserve(postfix.size());
+
+    auto add_sub_expr = [](std::vector<logic_int>& acc,
+                           std::vector<logic_int>& expr,
+                           bool parentheses) {
+        if (parentheses)
+        {
+            acc.push_back(logic::lopen);
+        }
+        acc.insert(acc.end(), expr.begin(), expr.end());
+        if (parentheses)
+        {
+            acc.push_back(logic::lclose);
+        }
+    };
+
+    // Process each token
+    for (auto lgc : postfix)
+    {
+        if (logic::is_operator_token(lgc))
+        {
+            switch (lgc)
+            {
+                case logic::ltrue:
+                    infix.push_back({logic::ltrue, {lgc}});
+                    break;
+                case logic::lor:
+                    [[fallthrough]];
+                case logic::land: {
+                    CELER_EXPECT(infix.size() > 1);
+                    auto& op_2 = infix.back();
+                    auto& op_1 = *(infix.end() - 2);
+                    auto opposite = lgc == logic::lor ? logic::land
+                                                      : logic::lor;
+                    std::vector<logic_int> new_expr;
+                    constexpr int max_extra_tokens = 5;
+                    new_expr.reserve(max_extra_tokens + op_1.expr.size()
+                                     + op_2.expr.size());
+                    add_sub_expr(
+                        new_expr, op_1.expr, op_1.expr_type == opposite);
+                    new_expr.push_back(lgc);
+                    add_sub_expr(
+                        new_expr, op_2.expr, op_2.expr_type == opposite);
+
+                    infix.pop_back();
+                    infix.pop_back();
+                    infix.push_back({logic::OperatorToken{lgc}, new_expr});
+                    break;
+                }
+                case logic::lnot: {
+                    CELER_EXPECT(!infix.empty());
+                    auto&& [expr_type, expr] = infix.back();
+                    std::vector<logic_int> new_expr;
+                    constexpr int max_extra_tokens = 3;
+                    new_expr.reserve(max_extra_tokens + expr.size());
+
+                    new_expr.push_back(lgc);
+                    add_sub_expr(new_expr, expr, expr_type < logic::lnot);
+
+                    infix.pop_back();
+                    infix.push_back({logic::lnot, new_expr});
+                    break;
+                }
+                default:
+                    CELER_ASSERT_UNREACHABLE();
+            }
+        }
+        else
+        {
+            infix.push_back({logic::ltrue, {lgc}});
+        }
+    }
+    CELER_ENSURE(infix.size() == 1);
+    return infix.front().expr;
+}
+
 //---------------------------------------------------------------------------//
 /*!
  * Result of building a logic representation of a node.
