+++ test/orange/CMakeLists.txt
@@ -84,6 +84,8 @@ celeritas_add_test(orangeinp/detail/BoundingZone.test.cc)
 celeritas_add_test(orangeinp/detail/PolygonUtils.test.cc)
 celeritas_add_test(orangeinp/detail/ProtoMap.test.cc)
 celeritas_add_test(orangeinp/detail/LocalSurfaceInserter.test.cc)
+celeritas_add_test(orangeinp/detail/LogicUtils.test.cc
+  LINK_LIBRARIES nlohmann_json::nlohmann_json)
 celeritas_add_test(orangeinp/detail/SurfaceGridHash.test.cc)
 celeritas_add_test(orangeinp/detail/TransformInserter.test.cc)
 
@@ -137,9 +139,6 @@ celeritas_add_test(univ/detail/LogicStack.test.cc)
 celeritas_add_test(univ/detail/RaggedRightIndexer.test.cc)
 celeritas_add_test(univ/detail/SurfaceFunctors.test.cc)
 celeritas_add_test(univ/detail/SenseCalculator.test.cc)
-celeritas_add_test(univ/detail/Utils.test.cc
-  LINK_LIBRARIES nlohmann_json::nlohmann_json)
-
 
 #-----------------------------------------------------------------------------#
 # Geant4 construction
